// ui/src/models/apiError.model.ts
/**
 * @license
 * Copyright 2021 Red Hat
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

// tslint:disable-next-line:interface-name
export interface ApiError {
    causes: RuleViolationCause[]|null
    message: string
    error_code: number
    detail: string
    name: string
}

// tslint:disable-next-line:interface-name
export interface RuleViolationCause {
    description: string
    context: string|null
}